requirements.txt
requests==2.31.0
python-dotenv==1.0.0
atlassian-python-api==3.41.0
json5==0.9.14
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
typer==0.9.0
rich==13.7.0
